another bunch of changes

+++ app/Http/Controllers/BlogController.php
@@ -4,8 +4,6 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Blog;
-use App\Models\Image;
-use App\Models\Comment;
 
 class BlogController extends Controller
 {
@@ -18,18 +16,27 @@ class BlogController extends Controller
 
     public function show(Blog $blog)
     {
-        // Load associated images and comments
-        $blog->load('images', 'comments');
-
-        return view('blogs.show', compact('blog'));
+        // Load associated images, comments, and the user relationship
+        $blog->load('images', 'comments', 'user');
+    
+        // Check if the user is authenticated
+        $authenticatedUserId = auth()->user() ? auth()->user()->id : null;
+    
+        // Use the "comments" relationship directly without calling the "get" method
+        $comments = $blog->comments;
+    
+        return view('blogs.show', compact('blog', 'comments', 'authenticatedUserId'));
     }
+    
+          
 
     public function create()
     {
         return view('blogs.create');
     }
 
-    public function test(){
+    public function test()
+    {
         return view('welcome');
     }
 
@@ -40,53 +47,63 @@ class BlogController extends Controller
             'content' => 'required|string',
             'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
         ]);
-
+    
         $blog = Blog::create([
             'title' => $request->input('title'),
             'content' => $request->input('content'),
+            'user_id' => auth()->user()->id, // Associate the user with the blog post
         ]);
-
+    
         // Handle Image Upload
         $imagePath = $request->file('image')->store('images');
         $blog->images()->create(['path' => $imagePath]);
-
+    
         return redirect()->route('blogs.index');
     }
+    
+    
 
-    public function edit($id)
+    public function edit(Blog $blog)
     {
-        $blog = Blog::findOrFail($id);
-        return view('blogs.edit', ['blog' => $blog]);
+        // Authorize the user to edit the blog
+        $this->authorize('update', $blog);
+
+        return view('blogs.edit', compact('blog'));
     }
 
-    public function update(Request $request, $id)
+    public function update(Request $request, Blog $blog)
     {
+        // Authorize the user to update the blog
+        $this->authorize('update', $blog);
+
         $request->validate([
             'title' => 'required|string|max:255',
             'content' => 'required|string',
         ]);
 
-        $blog = Blog::findOrFail($id);
         $blog->update([
             'title' => $request->input('title'),
             'content' => $request->input('content'),
         ]);
 
-        return redirect()->route('blogs.index');
+        return redirect()->route('blogs.show', $blog)->with('success', 'Blog updated successfully!');
     }
 
-    public function destroy($id)
+    public function destroy(Blog $blog)
     {
-        $blog = Blog::findOrFail($id);
+        // Authorize the user to delete the blog
+        $this->authorize('delete', $blog);
+
         $blog->delete();
 
-        return redirect()->route('blogs.index');
+        return redirect()->route('blogs.index')->with('success', 'Blog deleted successfully!');
     }
 
     public function __construct()
     {
         $this->middleware('auth')->except(['index', 'show']);
     }
+
 }
